+++ src/main/java/com/researchspace/dataverse/entities/DatasetFileDetails.java
@@ -1,9 +1,14 @@
 package com.researchspace.dataverse.entities;
 
+import com.researchspace.dataverse.http.FileUploadMetadata;
 import lombok.Data;
 
 import java.util.Date;
 
+/**
+ * DatasetFileDetails is a subsection of the response from
+ * {@link com.researchspace.dataverse.api.v1.DatasetOperations#uploadNativeFile(byte[], FileUploadMetadata, Identifier, String)}
+ */
 @Data
 public class DatasetFileDetails {
     private int id;