+++ src/main/java/com/researchspace/dataverse/entities/DatasetFileList.java
@@ -3,6 +3,10 @@ package com.researchspace.dataverse.entities;
 import lombok.Data;
 
 import java.util.List;
+
+/**
+ * DatasetFileList is the response from uploading a file using the native API
+ */
 @Data
 public class DatasetFileList {
     List<DatasetFile> files;