+++ src/main/java/com/researchspace/dataverse/http/FileUploadMetadata.java
@@ -10,7 +10,6 @@ import java.util.List;
  */
 @Data
 @Builder
-
 public class FileUploadMetadata {
     private String description;
     private String directoryLabel;