+++ src/main/java/com/researchspace/dataverse/http/NativeFileUploader.java
@@ -1,15 +1,19 @@
 package com.researchspace.dataverse.http;
 
-import org.springframework.core.io.ByteArrayResource;
-import org.springframework.http.*;
+import org.springframework.core.io.AbstractResource;
+import org.springframework.http.HttpEntity;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.MediaType;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
-import org.springframework.web.client.RestTemplate;
 
+/**
+ * NativeFileUploader is  helper class that performs upload of files using native API
+ */
 public class NativeFileUploader {
 
-    public HttpEntity<MultiValueMap<String, Object>> uploadFile(FileUploadMetadata metadata, String apiKey, byte[] data, String fName){
-        RestTemplate restTemplate = new RestTemplate();
+    public HttpEntity<MultiValueMap<String, Object>> createFileUploadEntity(FileUploadMetadata metadata, String apiKey, AbstractResource resource){
+
         MultiValueMap<String,Object> multipartRequest = new LinkedMultiValueMap<>();
 
         HttpHeaders requestHeaders = new HttpHeaders();
@@ -17,15 +21,8 @@ public class NativeFileUploader {
         requestHeaders.setContentType(MediaType.MULTIPART_FORM_DATA);//Main request's headers
 
         HttpHeaders requestHeadersAttachment = new HttpHeaders();
-        //requestHeadersAttachment.setContentType(MediaType.APPLICATION_OCTET_STREAM);// extract mediatype from file extensionHttpEntity<ByteArrayResource> attachmentPart;
-        ByteArrayResource fileAsResource = new ByteArrayResource(data){
-            @Override
-            public String getFilename(){
-                return fName;
-            }
-        };
-        HttpEntity attachmentPart = new HttpEntity<>(fileAsResource,requestHeadersAttachment);
 
+        HttpEntity attachmentPart = new HttpEntity<>(resource,requestHeadersAttachment);
         multipartRequest.set("file",attachmentPart);
 
         HttpHeaders requestHeadersJSON = new HttpHeaders();