+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 Significant changes since 0.1.0
 
+1.1.0 In progress
+
+- feature: Support upload of files to a dataset using native API. #16
+- feature: After creating a Dataset, the persistent ID is stored in the Identifier object. #22
+
 1.0.0 2022-11-21
 
 Increasing major version due to major updates to dependencies. However, there are no
@@ -8,7 +13,7 @@ breaking API changes in this library.
 - dependencies: Major dependency updates to Spring 5.3, Lombok 18.24. 
 - build: fix integration tests
 - build: enable integration test running through Github actions
-- build: test build and test on Java 8, 11, and 17
+- build: test build and test on Java 8, 11, and 17.
 
 0.2.0 2022-11-20